refactor: Double을 BigDecimal로 변경

+++ src/main/kotlin/blackjack/domain/DealerResult.kt
@@ -1,6 +1,8 @@
 package blackjack.domain
 
+import java.math.BigDecimal
+
 data class DealerResult(
     val dealer: Dealer,
-    val profit: Double,
+    val profit: BigDecimal,
 )
